+++ src/zhinst/utils/exceptions.py
@@ -1,0 +1,10 @@
+"""Package exceptions."""
+
+
+class CompatibilityError(RuntimeError):
+    """Compatibility error.
+
+    .. versionadded:: 0.3
+    """
+
+    ...